BFS was hanging here because the compare_and_swap template was comparing old_val with the value of loc (the pointer it self).
Passing loc as first argument makes the compiler instantiate compare_and_swap
with T=int * but then the function call __sync_bool_compare_and_swap as:

__sync_bool_compare_and_swap(&x, old_val, new_val);

This means that the type of the first argument is int ** which is wrong.

There is one more issue still to address. The family of functions __sync_*
works only with primitive types and templates in platform_atomics.h do not
address this issue.