Preliminar support to .graph (metis) format.
Metis format is used by the dimacs 10th challenge.  It may be useful
to support it.

The current implementation supports only a subset of the metis format
(no weight on edges or vertices). It seems to be buggy when the loaded
graph has repeated edges or vertices without any edge.